+++ main.tf
@@ -28,8 +28,11 @@ locals {
 
   secret_webhook_key = local.has_secrets ? var.atlantis_gitlab_user_token != "" ? "ATLANTIS_GITLAB_WEBHOOK_SECRET" : var.atlantis_github_user_token != "" ? "ATLANTIS_GH_WEBHOOK_SECRET" : "ATLANTIS_BITBUCKET_WEBHOOK_SECRET" : "unknown_secret_webhook_key"
 
+  # determine if the alb has authentication enabled, otherwise forward the traffic unauthenticated
+  alb_authenication_method = length(keys(var.alb_authenticate_oidc)) > 0 ? "authenticate-oidc" : length(keys(var.alb_authenticate_cognito)) > 0 ? "authenticate-cognito" : "forward"
+
   # Container definitions
-  container_definitions = var.custom_container_definitions == "" ? var.atlantis_bitbucket_user_token != "" ? module.container_definition_bitbucket.json : module.container_definition_github_gitlab.json : var.custom_container_definitions
+  container_definitions = var.custom_container_definitions == "" ? var.atlantis_bitbucket_user_token != "" ? module.container_definition_bitbucket.json_map_encoded_list : module.container_definition_github_gitlab.json_map_encoded_list : var.custom_container_definitions
 
   container_definition_environment = [
     {
@@ -72,6 +75,10 @@ locals {
       name  = "ATLANTIS_REPO_WHITELIST"
       value = join(",", var.atlantis_repo_whitelist)
     },
+    {
+      name  = "ATLANTIS_HIDE_PREV_PLAN_COMMENTS"
+      value = var.atlantis_hide_prev_plan_comments
+    },
   ]
 
   # Secret access tokens
@@ -99,8 +106,6 @@ locals {
 }
 
 data "aws_region" "current" {}
-
-data "aws_caller_identity" "current" {}
 
 data "aws_route53_zone" "this" {
   count = var.create_route53_record ? 1 : 0
@@ -161,7 +166,7 @@ resource "aws_ssm_parameter" "atlantis_bitbucket_user_token" {
 ###################
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "v2.33.0"
+  version = "v2.47.0"
 
   create_vpc = var.vpc_id == ""
 
@@ -183,7 +188,7 @@ module "vpc" {
 ###################
 module "alb" {
   source  = "terraform-aws-modules/alb/aws"
-  version = "v5.5.0"
+  version = "v5.7.0"
 
   name     = var.name
   internal = var.internal
@@ -200,12 +205,13 @@ module "alb" {
 
   https_listeners = [
     {
-      target_group_index = 0
-      port               = 443
-      protocol           = "HTTPS"
-      certificate_arn    = var.certificate_arn == "" ? module.acm.this_acm_certificate_arn : var.certificate_arn
-      action_type        = length(keys(var.alb_authenticate_oidc)) > 0 ? "authenticate-oidc" : "forward"
-      authenticate_oidc  = var.alb_authenticate_oidc
+      target_group_index   = 0
+      port                 = 443
+      protocol             = "HTTPS"
+      certificate_arn      = var.certificate_arn == "" ? module.acm.this_acm_certificate_arn : var.certificate_arn
+      action_type          = local.alb_authenication_method
+      authenticate_oidc    = var.alb_authenticate_oidc
+      authenticate_cognito = var.alb_authenticate_cognito
     },
   ]
 
@@ -259,7 +265,7 @@ resource "aws_lb_listener_rule" "unauthenticated_access_for_cidr_blocks" {
 ###################
 module "alb_https_sg" {
   source  = "terraform-aws-modules/security-group/aws//modules/https-443"
-  version = "v3.9.0"
+  version = "v3.15.0"
 
   create = var.alb_https_security_group_id == "" ? true : false
 
@@ -267,14 +273,14 @@ module "alb_https_sg" {
   vpc_id      = local.vpc_id
   description = "Security group with HTTPS ports open for specific IPv4 CIDR block (or everybody), egress ports are all world open"
 
-  ingress_cidr_blocks = var.alb_ingress_cidr_blocks
+  ingress_cidr_blocks = sort(compact(concat(var.allow_github_webhooks ? var.github_webhooks_cidr_blocks : [], var.alb_ingress_cidr_blocks)))
 
   tags = local.tags
 }
 
 module "alb_http_sg" {
   source  = "terraform-aws-modules/security-group/aws//modules/http-80"
-  version = "v3.9.0"
+  version = "v3.15.0"
 
   create = var.alb_http_security_group_id == "" ? true : false
 
@@ -282,14 +288,14 @@ module "alb_http_sg" {
   vpc_id      = local.vpc_id
   description = "Security group with HTTP ports open for specific IPv4 CIDR block (or everybody), egress ports are all world open"
 
-  ingress_cidr_blocks = var.alb_ingress_cidr_blocks
+  ingress_cidr_blocks = sort(compact(concat(var.allow_github_webhooks ? var.github_webhooks_cidr_blocks : [], var.alb_ingress_cidr_blocks)))
 
   tags = local.tags
 }
 
 module "atlantis_sg" {
   source  = "terraform-aws-modules/security-group/aws"
-  version = "v3.9.0"
+  version = "v3.15.0"
 
   create = var.atlantis_security_group_id == "" ? true : false
 
@@ -317,7 +323,7 @@ module "atlantis_sg" {
 ###################
 module "acm" {
   source  = "terraform-aws-modules/acm/aws"
-  version = "v2.5.0"
+  version = "v2.10.0"
 
   create_certificate = var.certificate_arn == ""
 
@@ -350,9 +356,10 @@ resource "aws_route53_record" "atlantis" {
 ###################
 module "ecs" {
   source  = "terraform-aws-modules/ecs/aws"
-  version = "v2.0.0"
+  version = "v2.3.0"
 
-  name = var.name
+  name               = var.name
+  container_insights = var.ecs_container_insights
 
   tags = local.tags
 }
@@ -391,12 +398,11 @@ data "aws_iam_policy_document" "ecs_task_access_secrets" {
   statement {
     effect = "Allow"
 
-    resources = [
-      "arn:${var.aws_ssm_path}:ssm:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:parameter${var.webhook_ssm_parameter_name}",
-      "arn:${var.aws_ssm_path}:ssm:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:parameter${var.atlantis_github_user_token_ssm_parameter_name}",
-      "arn:${var.aws_ssm_path}:ssm:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:parameter${var.atlantis_gitlab_user_token_ssm_parameter_name}",
-      "arn:${var.aws_ssm_path}:ssm:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:parameter${var.atlantis_bitbucket_user_token_ssm_parameter_name}",
-    ]
+    resources = flatten(list(
+      aws_ssm_parameter.webhook.*.arn,
+      aws_ssm_parameter.atlantis_github_user_token.*.arn,
+      aws_ssm_parameter.atlantis_gitlab_user_token.*.arn,
+    aws_ssm_parameter.atlantis_bitbucket_user_token.*.arn))
 
     actions = [
       "ssm:GetParameters",
@@ -438,7 +444,7 @@ resource "aws_iam_role_policy" "ecs_task_access_secrets" {
 
 module "container_definition_github_gitlab" {
   source  = "cloudposse/ecs-container-definition/aws"
-  version = "v0.23.0"
+  version = "v0.40.0"
 
   container_name  = var.name
   container_image = local.atlantis_image
@@ -446,6 +452,21 @@ module "container_definition_github_gitlab" {
   container_cpu                = var.ecs_task_cpu
   container_memory             = var.ecs_task_memory
   container_memory_reservation = var.container_memory_reservation
+
+  user                     = var.user
+  ulimits                  = var.ulimits
+  entrypoint               = var.entrypoint
+  command                  = var.command
+  working_directory        = var.working_directory
+  repository_credentials   = var.repository_credentials
+  docker_labels            = var.docker_labels
+  start_timeout            = var.start_timeout
+  stop_timeout             = var.stop_timeout
+  container_depends_on     = var.container_depends_on
+  essential                = var.essential
+  readonly_root_filesystem = var.readonly_root_filesystem
+  mount_points             = var.mount_points
+  volumes_from             = var.volumes_from
 
   port_mappings = [
     {
@@ -464,6 +485,7 @@ module "container_definition_github_gitlab" {
     }
     secretOptions = []
   }
+  firelens_configuration = var.firelens_configuration
 
   environment = concat(
     local.container_definition_environment,
@@ -479,7 +501,7 @@ module "container_definition_github_gitlab" {
 
 module "container_definition_bitbucket" {
   source  = "cloudposse/ecs-container-definition/aws"
-  version = "v0.23.0"
+  version = "v0.40.0"
 
   container_name  = var.name
   container_image = local.atlantis_image
@@ -487,6 +509,21 @@ module "container_definition_bitbucket" {
   container_cpu                = var.ecs_task_cpu
   container_memory             = var.ecs_task_memory
   container_memory_reservation = var.container_memory_reservation
+
+  user                     = var.user
+  ulimits                  = var.ulimits
+  entrypoint               = var.entrypoint
+  command                  = var.command
+  working_directory        = var.working_directory
+  repository_credentials   = var.repository_credentials
+  docker_labels            = var.docker_labels
+  start_timeout            = var.start_timeout
+  stop_timeout             = var.stop_timeout
+  container_depends_on     = var.container_depends_on
+  essential                = var.essential
+  readonly_root_filesystem = var.readonly_root_filesystem
+  mount_points             = var.mount_points
+  volumes_from             = var.volumes_from
 
   port_mappings = [
     {
@@ -505,6 +542,7 @@ module "container_definition_bitbucket" {
     }
     secretOptions = []
   }
+  firelens_configuration = var.firelens_configuration
 
   environment = concat(
     local.container_definition_environment,
@@ -560,6 +598,8 @@ resource "aws_ecs_service" "atlantis" {
     container_port   = var.atlantis_port
     target_group_arn = element(module.alb.target_group_arns, 0)
   }
+
+  tags = local.tags
 }
 
 ###################